extract base class from pipeline implementations
